Wrote a slightly more meaningful humans.txt
It's Christmas Eve. I'm wearing my favorite shoes but yet I feel sad. Why? It's probably the headaches, or the fact that I am here in Miami.
It's so weird I always feel like this every time I come by now.

In this time of appreciation, written messages persist forever, and what best way to express my gratefulness to those who have been part of
my life than to write them a sort-of-open love letter?
Previously there was a humans.txt, but it was simply a list of people alongside the time I've met them. Now, I took some time to write some
of what I think about them, so if someone ever stumbles upon this file, they'll know why is it there for.

I guess now what is left is to survive the night with a enormous amount of noise pollution around.
Happy Holidays!